grub signing integration: Add support for .sig files and grub signing
This commit allows ostree to work fully with using a gpg certificate
which gets built into grub for verifying all the loaded files,
include the grub.cfg file.  The following variables have to be set in
order to make use of the new functionality.

GRUB_SIGN_VERIFY = "1"
UEFI_SELOADER = "0"

The password in the grub.cfg file is now removed from the build
configuration files and stored in a file.  The script provided in
meta-secure-core/meta-signing-key/scripts/create-user-key-store.sh can
generate a new key and provides a keys.conf which can be included in
the local.conf to make use of non-sample keys.

[ Issue: LINUXEXEC-5303 ]

+++ classes/image_types_ota.bbclass
@@ -96,7 +96,12 @@ IMAGE_CMD_otaimg () {
 			mkdir -p ${PHYS_SYSROOT}/boot/efi/EFI/BOOT
 			if [ -n "${@bb.utils.contains('DISTRO_FEATURES', 'efi-secure-boot', 'Y', '', d)}" ]; then
 				cp ${DEPLOY_DIR_IMAGE}/grubx64.efi ${PHYS_SYSROOT}/boot/efi/EFI/BOOT/bootx64.efi
-				cp ${DEPLOY_DIR_IMAGE}/grub.cfg.p7b ${PHYS_SYSROOT}/boot/efi/EFI/BOOT/
+				if [ -f ${DEPLOY_DIR_IMAGE}/grub.cfg.p7b ] ; then
+					cp ${DEPLOY_DIR_IMAGE}/grub.cfg.p7b ${PHYS_SYSROOT}/boot/efi/EFI/BOOT/
+				fi
+				if [ -f ${DEPLOY_DIR_IMAGE}/grub.cfg.sig ] ; then
+					cp ${DEPLOY_DIR_IMAGE}/grub.cfg.sig ${PHYS_SYSROOT}/boot/efi/EFI/BOOT/
+				fi
 			else
 				cp ${DEPLOY_DIR_IMAGE}/grub-efi-bootx64.efi ${PHYS_SYSROOT}/boot/efi/EFI/BOOT/bootx64.efi
 			fi
